# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rechvix API
  version: "0.1.0"
  description: >
    Multi-company, multi-branch, multi-warehouse billing, inventory,
    accounting, and India-GST tax-management API. This spec is generated
    by reading `internal/modules/*/httpapi/handlers.go` directly — every
    path, method, and request/response shape below is verified against
    the actual Go handler code, not inferred from `docs/architecture.md`'s
    design intent. See `docs/api.md` for narrative conventions (auth,
    error envelope, money shape, pagination status) that apply across
    every endpoint below rather than being repeated per-operation.

    Status: complete — all 124 real routes across all 15 HTTP modules are
    present and typed, mechanically verified against a grep of every
    r.Get/Post/Put/Delete/Patch(...) registration in the codebase (exact
    match, see the git history for this file). Kept current as endpoints
    ship (docs/TODO.md's cross-cutting checklist item) — re-run that same
    grep diff after adding or changing a route, don't just trust this
    note.
servers:
  - url: /api/v1
    description: Same origin as the web app; also reachable directly on the configured HTTP_PORT.
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /pricing/price-lists:
    get:
      tags: [Pricing]
      summary: List price lists.
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { price_lists: { type: array, items: { $ref: "#/components/schemas/PriceList" } } } } } } }
    post:
      tags: [Pricing]
      summary: Create a price list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, currency_code]
              properties:
                name: { type: string }
                currency_code: { type: string }
                is_default: { type: boolean }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PriceList" } } } }
  /pricing/price-lists/{id}:
    get:
      tags: [Pricing]
      summary: Get a price list.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PriceList" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /pricing/price-lists/{id}/items:
    get:
      tags: [Pricing]
      summary: List price items in a price list.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/PriceListItem" } } } } } } }
    post:
      tags: [Pricing]
      summary: Set the price of a variant/unit combination in this price list.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_variant_id, unit_id, amount, currency_code]
              properties:
                product_variant_id: { type: string, format: uuid }
                unit_id: { type: string, format: uuid }
                amount: { type: string }
                currency_code: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PriceListItem" } } } }
        "400":
          description: NEGATIVE_PRICE / INVALID_AMOUNT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /pricing/price-lists/{id}/resolve:
    get:
      tags: [Pricing]
      summary: Resolve the effective price of a variant/unit in this price list (the Sales billing screen's price lookup).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: variant_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: unit_id, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PriceListItem" } } } }
        "404": { $ref: "#/components/responses/NotFound" }

  # ============================================================
  # INVENTORY — append-only stock movements, materialized balances,
  # transfers, adjustments, reservations, low-stock, policies
  # Verified against internal/modules/inventory/domain/domain.go.
  # ============================================================
...
  /purchases/documents:
    get:
      tags: [Purchases]
      summary: List purchase documents, optionally filtered by document_type.
      parameters:
        - { name: document_type, in: query, schema: { $ref: "#/components/schemas/PurchaseDocumentType" } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { documents: { type: array, items: { $ref: "#/components/schemas/PurchaseDocument" } } } } } } }
    post:
      tags: [Purchases]
      summary: Create a purchase document (DRAFT).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_id, warehouse_id, supplier_party_id, document_type, currency_code]
              properties:
                branch_id: { type: string, format: uuid }
                warehouse_id: { type: string, format: uuid }
                supplier_party_id: { type: string, format: uuid }
                document_type: { $ref: "#/components/schemas/PurchaseDocumentType" }
                reference_document_id: { type: string, format: uuid, nullable: true, description: e.g. the PO a GRN is receiving against. }
                supplier_invoice_number: { type: string }
                supplier_invoice_date: { type: string, format: date-time, nullable: true }
                document_date: { type: string, format: date-time, nullable: true }
                currency_code: { type: string }
                notes: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PurchaseDocument" } } } }
        "400":
          description: INVALID_DOCUMENT_TYPE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /purchases/documents/{id}:
    get:
      tags: [Purchases]
      summary: Get a purchase document with its lines.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  document: { $ref: "#/components/schemas/PurchaseDocument" }
                  lines: { type: array, items: { $ref: "#/components/schemas/PurchaseDocumentLine" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /purchases/documents/{id}/lines:
    post:
      tags: [Purchases]
      summary: Add a line to a DRAFT purchase document.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_variant_id, unit_id, quantity, unit_price]
              properties:
                product_variant_id: { type: string, format: uuid }
                unit_id: { type: string, format: uuid }
                quantity: { type: string }
                unit_price: { type: string }
                batch_code: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PurchaseDocumentLine" } } } }
        "409":
          description: DOCUMENT_NOT_DRAFT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /purchases/documents/{id}/finalize:
    post:
      tags: [Purchases]
      summary: Finalize a purchase document — posts stock (for stock-affecting types) and, since Stage 6, auto-posts the accounting journal. Immutable afterward.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PurchaseDocument" } } } }
        "409":
          description: DOCUMENT_NOT_DRAFT / EMPTY_DOCUMENT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  # ============================================================
  # GST (India) — admin CRUD for tax_rate_master only. Real tax
  # calculation is server-internal (taxation.CalculateAndSnapshotTx,
  # invoked by the sales module on finalize) — not a callable endpoint,
  # since a client-computed tax figure is never trusted (docs/api.md).
  # Verified against internal/modules/gstindia/domain/domain.go.
  # ============================================================
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bp_session
      description: Set by POST /auth/login. Default cookie name; configurable via SESSION_COOKIE_NAME.
    apiKey:
      type: http
      scheme: bearer
      description: >
        Authorization: Bearer <key>. Effective permissions are the
        intersection of the key's declared scopes and the owning user's
        real RBAC grants — see docs/api.md.
  responses:
...
    NotFound:
      description: Resource does not exist, or exists in a different organisation (never distinguished from the client's perspective).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: SESSION_INVALID }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
    Money:
      type: object
      description: >
        amount is a STRING, always — see docs/api.md. Never parse it as a
        JS/JSON number for anything beyond display formatting.
      required: [amount, currency]
      properties:
        amount: { type: string, example: "1234.50" }
        currency: { type: string, example: INR }
...
    PriceList:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        Name: { type: string }
        CurrencyCode: { type: string }
        IsDefault: { type: boolean }
        Status: { type: string }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    PriceListItem:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        PriceListID: { type: string, format: uuid }
        ProductVariantID: { type: string, format: uuid }
        UnitID: { type: string, format: uuid }
        Price: { $ref: "#/components/schemas/Money" }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
...
    PurchaseDocumentType:
      type: string
      enum: [PURCHASE_ORDER, GOODS_RECEIPT, PURCHASE_INVOICE, PURCHASE_RETURN, DEBIT_NOTE]
    PurchaseDocument:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        BranchID: { type: string, format: uuid }
        WarehouseID: { type: string, format: uuid }
        SupplierPartyID: { type: string, format: uuid }
        DocumentType: { $ref: "#/components/schemas/PurchaseDocumentType" }
        DocumentNumber: { type: string }
        ReferenceDocumentID: { type: string, format: uuid, nullable: true }
        Status: { type: string, enum: [DRAFT, FINALIZED, CANCELLED] }
        SupplierInvoiceNumber: { type: string }
        SupplierInvoiceDate: { type: string, format: date-time, nullable: true }
        DocumentDate: { type: string, format: date-time }
        CurrencyCode: { type: string }
        Notes: { type: string }
        CreatedBy: { type: string, format: uuid }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        FinalizedAt: { type: string, format: date-time, nullable: true }
    PurchaseDocumentLine:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        PurchaseDocumentID: { type: string, format: uuid }
        LineNumber: { type: integer }
        ProductVariantID: { type: string, format: uuid }
        UnitID: { type: string, format: uuid }
        Quantity: { type: string }
        UnitPrice: { $ref: "#/components/schemas/Money" }
        LineTotal: { $ref: "#/components/schemas/Money" }
        BatchCode: { type: string }
        CreatedAt: { type: string, format: date-time }
...